+++ functions/index.js
@@ -14,7 +14,6 @@ admin.initializeApp();
 const { handleProgressionUpdates } = require('./updateProgression');
 const { automatedBackups, automatedRestore } = require('./backupAndRestore');
 const { scheduledFunction, triggerScheduledFunction } = require('./automatedKeyResults');
-
 /*
 |----------------------------------------------------------------------------------
 |
@@ -47,3 +46,9 @@ exports.automatedRestore = automatedRestore();
  */
 exports.scheduledFunction = scheduledFunction();
 exports.triggerScheduledFunction = triggerScheduledFunction();
+
+/**
+ * Function for populating mock data to Firebase emulator when developing
+ * locally. Will not be used in production.
+ */
+exports.populateFirestoreEmulator = require('./populateFirestoreEmulator');