+++ src/components/DocumentSidebar.vue
@@ -7,6 +7,7 @@
             name: editLinkName,
             params: { slug: $route.params.slug },
           }"
+          data-cy="edit_object_link"
           class="sidebar-nav__item"
           v-tooltip.right="$t('tooltip.changeDetailsAndTeam')"
         >
@@ -18,6 +19,7 @@
             class="sidebar-nav__item"
             @click="expandAddObjective = true"
             v-tooltip.right="$t('tooltip.addObjective')"
+            data-cy="add_objective"
           >
             <i class="fa fas fa-fw fa-plus"></i>
             {{ $t('document.addObjective') }}
@@ -35,6 +37,7 @@
             class="sidebar-nav__item"
             @click="expandAddKeyRes = true"
             v-tooltip.right="$t('tooltip.addKeyresToObjective')"
+            data-cy="add_keyres"
           >
             <i class="fa fas fa-fw fa-plus"></i>
             {{ $t('document.newKeyres') }}
@@ -51,6 +54,7 @@
             class="sidebar-nav__item"
             @click="expandRegisterProgress = true"
             v-tooltip.right="$t('tooltip.updateKeyres')"
+            data-cy="update_data"
           >
             <i class="fa fas fa-fw fa-chart-line"></i>
             {{ $t('document.updateData') }}