+++ src/views/Admin/AdminHome.vue
@@ -6,7 +6,7 @@
       <div class="container">
         <nav class="nav">
           <router-link class="router-link" :to="{ name: 'admin-users' }">{{ $t('admin.home.users') }}</router-link>
-          <router-link class="router-link" :to="{ name: 'admin-objects' }">{{
+          <router-link class="router-link" :to="{ name: 'admin-objects' }" data-cy="admin-depsAndProds">{{
             $t('admin.home.depsAndProds')
           }}</router-link>
         </nav>