+++ CHANGELOG.md
@@ -4,14 +4,23 @@ All notable changes to this project will be documented in this file. The format 
 
 ## Unreleased
 
+### Added
+
+- e2e tests with Cypress. See [documentation](/tests/e2e/README.md) and [test plan](/tests/e2e/test-plan.md).
+- Local development with Firebase Emulator and mock data:
+  - Use `npm run local` to run Emulator Suite, development server and inject mock data til emulated Firestore
+
 ### Changed
 
+- Improve documentation
 - Update prettier to version 2.0
 - Update dependencies
 
 ### Fixed
 
 - Typo in documentation
+- Fix bug in manual testing of automatic key results (trigger cloud function) in UI with better feedback and error handling
+- Gitignore `.firebaserc`
 
 ## [1.4.2] - 2020-04-22